pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deepwiki2md"
version = "1.0.0"
description = "A modern Python package for scraping DeepWiki content to Markdown using PyDoll"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "deepwiki2md", email = "contact@deepwiki2md.dev"}
]
keywords = ["scraping", "markdown", "deepwiki", "documentation", "pydoll"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers", 
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: Indexing/Search",
    "Topic :: Text Processing :: Markup :: HTML",
    "Topic :: Text Processing :: Markup :: Markdown",
]
requires-python = ">=3.8"
dependencies = [
    "pydoll-python>=2.0.0",
    "beautifulsoup4>=4.12.0",
    "markdownify>=0.11.6",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-asyncio>=0.21.0",
    "black>=22.0",
    "isort>=5.0",
    "flake8>=4.0",
    "mypy>=0.900",
]

[project.urls]
"Homepage" = "https://github.com/deepwiki2md/deepwiki2md"
"Bug Reports" = "https://github.com/deepwiki2md/deepwiki2md/issues"
"Source" = "https://github.com/deepwiki2md/deepwiki2md"

[project.scripts]
deepwiki2md = "deepwiki2md.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["deepwiki2md*"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"